Base

Description

Functionality for: * Converting decimal numbers to any base/radix * Converting numbers from any baseradix to decimal baseradix

Documentation

toBase :: Int -> Int -> [Int]

The toBase function converts a number from decimal base to a specified base in the form of digits. The function takes two arguments, both of type Int.

Below are a few examples:

>>> toBase 8 37
[4,5]

>>> toBase 100 233243
[23,32,43]

>>> toBase 9 233243
[3,8,4,8,4,8]

>>> toBase 35 233243
[5,15,14,3]

fromBase :: Int -> [Int] -> Int

The fromBase function converts a number from a specified base to decimal in the form of a list of digits. The function takes two arguments, the first of type Int and the second of type [Int].

Below are a few examples:

>>> fromBase 16 [15,15,15]
4095

>>> fromBase 100 [21,12,1]
211201

>>> fromBase 2 [1,0,1]
5

toAlphaDigits :: [Int] -> String

The toAlphaDigits function converts a number from a list of Int to corresponding String representation. The function takes one argument of type [Int], which contains the list of digits.

Below are a few examples:

>>> toAlphaDigits $ toBase 16 23432
"5b88"

>>> toAlphaDigits $ toBase 16 255
"ff"

>>> toAlphaDigits [38,12,1]
"}c1"

>>> toAlphaDigits [21,12,1]
"lc1"

fromAlphaDigits :: String -> [Int]

The 'from AlphaDigits' function converts a String to list of Int digits. The function takes only one argument of type String.

Below are a few examples:

>>> fromAlphaDigits "j43hbrh"
[19,4,3,17,11,27,17]

>>> fromAlphaDigits "ffff"
[15,15,15,15]