Combinatorics

Description

Functionality for: * Factorial * Finding number of combinations * Finding number of permutations * Generating permutations and combinations * Generating random permutation (shuffle)

Documentation

factorial :: Integer -> Integer

The factorial function returns the factorial of an integer. It takes one argument, of type Integer.

Three examples are given below:

>>> factorial 5
120

>>> factorial 25
15511210043330985984000000

>>> factorial (-1)
*** Exception: Usage - factorial n, where 'n' is non-negative.

c :: Integer -> Integer -> Integer

The c function returns the number of combinations. It takes two arguments, both of type Integer. When called as c n r, where n and r are arguments, it returns the number of possible combinations of r objects from a set of n objects.

Two examples are given below:

>>> c 10 5
252

>>> 18 `c` 8
43758

p :: Integer -> Integer -> Integer

The p function returns the number of permutations. It takes two arguments, both of type Integer. When called as p n r, where n and r are arguments, it returns the number of possible permutations of r objects from a set of n objects.

Two examples are given below:

>>> p 10 5
30240

>>> 18 `p` 8
1764322560

permutation :: [a] -> [[a]]

The permutation function returns all possible permutations of a list. It takes one argument, which can be a list or string.

Two examples are given below:

>>> permutation "abc"
["abc","bac","cba","bca","cab","acb"]

>>> permutation [1,2]
[[1,2],[2,1]]

shuffle :: [a] -> IO [a]

The shuffle function generates a random permutation. The function takes a list as argument and returns a shuffled list using the Fisher-Yates shuffle algorithm.

Six examples are given below:

>>> shuffle [1..10]
[6,10,1,7,4,8,9,3,2,5]

>>> shuffle [2,4..10]
[8,4,10,2,6]

>>> shuffle [2,3,5,7]
[3,7,5,2]

>>> shuffle ["abc", "xys"]
["abc","xys"]

>>> shuffle ["abc", "xys", "fdvdfv"]
["abc","fdvdfv","xys"]

>>> shuffle [[1,2], [3,4], [5,6]]
[[3,4],[1,2],[5,6]]

combination :: Applicative f => Int -> f [a] -> f [a]

The combination function generates all possible combinations. The function takes two arguments, the first of type Integer and the second as a list of lists. The function generates a combination of n elements at a time, where n is the first argument.

Three examples are given below:

>>> combination 2 ["a","b","c"]
["aa","ab","ac","ba","bb","bc","ca","cb","cc"]

>>> combination 2 [[1],[2],[3]]
[[1,1],[1,2],[1,3],[2,1],[2,2],[2,3],[3,1],[3,2],[3,3]]