Fibonacci

Description

Functionality for: * Calculating the nth term in the Fibonacci Series * Generating Fibonacci Series with n terms

Documentation

fib :: Integer -> Integer

The fib function returns the nth term of the Fibonacci sequence. The function takes one argument of type Integer, whose value should be positive.

Below are a few examples:

>>> fib 10
55

>>> fib 100
354224848179261915075

fibSeq :: Integer -> [Integer]

The fibSeq function returns the first n terms of the Fibonacci sequence. The function takes one argument of type Integer.

Below are a few examples:

>>> fibSeq 10
[1,1,2,3,5,8,13,21,34,55]

>>> fibSeq 20
[1,1,2,3,5,8,13,21,34,55,89,144,233,377,610,987,1597,2584,4181,6765]

fibIndex :: Integer -> Integer

The fibIndex function returns the index of the passed Fibonacci number in the Fibonacci sequence. The function takes one argument of type Integer.

Below are a few examples:

>>> fibIndex 144
12

>>> fibIndex 6765
20

>>> fibIndex (fib 10000)
10000

isFibNum :: Integer -> Bool

The isFibNum function checks whether an integer is a Fibonacci number. The function takes one argument of type Integer.

Below are a few examples:

>>> isFibNum 144
True

>>> isFibNum (144 + 200)
False

>>> isFibNum (fib 1000)
True