Graph

Description

Functionality for: * Data types (Graph, GraphMatrix, Edges and Vertices) * Adding edges and vertices * Getting vertices and their number * Getting edges and their number * Converting between different Graph data types * Transpose * Checking for directedness * Union * Connectedness of vertices * Number of paths between two vertices * Adjacent nodes of a vertex * Degrees of vertices * Euler circuits and paths * Hamiltonian circuits

Documentation

newtype Vertices a

Vertices is a data type for representing the vertices of a graph. It is the set of all the vertices in a graph. Vertices are internally stored as a list.

Below are a few examples:

>>> Vertices []
{}

>>> Vertices [1,2,5]
{1,2,5}

>>> Vertices [1..10]
    {1,2,3,4,5,6,7,8,9,10}

Constructors

Vertices [a]

newtype Edges a

Edges is a data types for representing the edges of a graph. It is the set of all edges in a graph, including their starting and ending vertices, and the weight of the edge.

Below are a few examples:

>>> Edges [(1,1,10)]
{(1,1,10)}

>>> Edges [(1,2,1),(1,3,5),(2,3,2)]
{(1,2,1),(1,3,5),(2,3,2)}

>>> Edges []
{}

Constructors

Edges [(a, a, Int)]

newtype Graph a

Graph is a data type for representing graphs as a group of Vertices and Edges.

For example:

>>> let v = Vertices [1,2,3,4]
>>> v
{1,2,3,4}

>>> let e = Edges [(1,2,5),(1,3,7),(2,4,3)]
>>> e
{(1,2,5),(1,3,7),(2,4,3)}

>>> let g = Graph (v,e)
>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

Constructors

Graph (Vertices a, Edges a)

newtype GraphMatrix a

The GraphMatrix data type is used here to represent a Graph as a two-dimentional list.

For example:

>>> GraphMatrix [[1,2,5],[3,5,7],[0,0,4]]
1 2 5
3 5 7
0 0 4

Constructors

GraphMatrix [[a]]

verticesToList :: Vertices t -> [t]

The verticesToList function returns the list of vertices of a graph represented as Vertices.

Below is an example:

>>> v1
{1,2,3,4}

>>> verticesToList v1
    [1,2,3,4]

listToVertices :: [t] -> Vertices t

The list2vertices function returns a Vertices from a list of vertices.

For example:

>>> list2vertices [1,2,3,4]
{1,2,3,4}

edgesToList :: Edges t -> [(t, t, Int)]

The edgesToList function returns the list of edges of a graph represented as Edges.

Below is an example:

>>> e1
{(1,2,4),(1,3,1),(1,4,5),(3,2,6)}

>>> edgesToList e1
[(1,2,4),(1,3,1),(1,4,5),(3,2,6)]

listToEdges :: [(a, a, Int)] -> Edges a

The listToEdges function converts a list of triples into Edges.

For example:

>>> l
[(1,2,1),(2,2,4)]

>>> listToEdges l
{(1,2,1),(2,2,4)}

graphToMatrix :: GraphMatrix a -> [[a]]

The graphToMatrix function converts GraphMatrix type to a two-dimentional list.

For example:

>>> gm1
0 1 3 5
4 0 2 6
1 2 3 4
0 0 1 6

>>> graphToMatrix gm1
[[0,1,3,5],[4,0,2,6],[1,2,3,4],[0,0,1,6]]

matrixToGraph :: [[a]] -> GraphMatrix a

The matrixToGraph function converts a two-dimensional list to a GraphMatrix.

For example:

>>> m
[[0,1],[4,0]]

>>> graphToMatrix m
0   1
4   0

getVerticesG :: Num a => Graph a -> Vertices a

The getVerticesG function returns the set of all vertices from a graph. The function takes one argument, which is of type Graph.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> getVerticesG g
{1,2,3,4}

getVerticesGM :: GraphMatrix t -> Vertices Int

The getVerticesGM returns the edges in a graph from its adjacency matrix representation. The function takes one argument, which is of type GraphMatrix.

For example:

>>> gm
    1	2	5
    3	5	7
    0	0	4

>>> getVerticesGM gm
{1,2,3}

numVerticesG :: Num a => Graph a -> Int

The numVerticesG function returns the number of vertices in a graph. The function takes one argument, which is of type Graph.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> numVerticesG g
4

numVerticesGM :: Num a => GraphMatrix a -> Int

The numVerticesGM function returns the number of vertices in a graph from its adjacency matrix representation. The function takes one argument, which is of type GraphMatrix.

For example:

>>> gm
1	2	5
3	5	7
0	0	4

>>> numVerticesGM gm
3

getEdgesG :: Num a => Graph a -> Edges a

The getEdgesG function returns the edges of a graph. The function takes one argument, which is of type Graph.

For example:

>>> g
    Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> getEdgesG g
{(1,2,5),(1,3,7),(2,4,3)}

getEdgesGM :: Integral a => GraphMatrix a -> Edges Int

The getEdgesGM function returns the set of edges in a graph from its adjacency matrix representation. The function takes one argument, which is of type GraphMatrix.

For example:

>>> gm
1	2	5
3	5	7
0	0	4

>>> getEdgesGM gm
{(1,1,1),(1,2,2),(1,3,5),(2,1,3),(2,2,5),(2,3,7),(3,3,4)}

numEdgesG :: Num a => Graph a -> Int

The numEdgesG function returns the number of edges in a graph. The function takes one argument, which is of type Graph.

For example:

>>> g
    Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> numEdgesG g
3

numEdgesGM :: Integral a => GraphMatrix a -> Int

The numEdgesGM function returns the number of edges in a graph from its adjacency matrix representation. The function takes one argument, which is of the type GraphMatrix.

For example:

>>> gm
1	2	5
3	5	7
0	0	4

>>> numEdgesGM gm
7

convertGM2G :: Integral a => GraphMatrix a -> Graph Int

The convertGM2G function returns a Graph representation as G(V,E) from its adjacency matrix representation. The function takes one argument, which is of type GraphMatrix and returns a Graph.

For example:

>>> gm
1	2	5
3	5	7
0	0	4

>>> convertGM2G gm
Graph ({1,2,3},{(1,1,1),(1,2,2),(1,3,5),(2,1,3),(2,2,5),(2,3,7),(3,3,4)})

convertG2GM :: (Eq t, Num t) => Graph t -> GraphMatrix Int

The convertG2GM function returns an adjacency matrix representation of a graph. The function takes one argument, which is of type Graph and returns a GraphMatrix.

For example:

>>> g
	Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> convertG2GM g
0	5	7	0
0	0	0	3
0	0	0	0
0	0	0	0

transposeG :: (Eq t, Num t) => Graph t -> Graph Int

The transposeG function returns a transpose of a graph. The function takes one argument, which is of type Graph and returns another graph of type Graph.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> transposeG g
Graph ({1,2,3,4},{(2,1,5),(3,1,7),(4,2,3)})

transposeGM :: Num a => GraphMatrix a -> GraphMatrix a

The transposeGM function returns a transpose of a graph (in adjacency matrix form). The function takes one argument, which is of type GraphMatrix and returns a GraphMatrix.

For example:

>>> gm
1	2	5
3	5	7
0	0	4

>>> transposeGM gm
1	3	0
2	5	0
5	7	4

isUndirectedG :: (Num a, Eq a) => Graph a -> Bool

The isUndirectedG function returns whether a graph is undirected or not. A graph is undirected if it is equal to its transpose. The function takes one argument, which is of type Graph.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> isUndirectedG g
False

isUndirectedGM :: (Num a, Eq a) => GraphMatrix a -> Bool

The isUndirectedGM function returns whether a graph (adjacency matrix form) is undirected or not. A graph is undirected if it is equal to its transpose. The function takes one argument, which is of type GraphMatrix.

Below are two examples:

>>> gm1
1	2	5
3	5	7
0	0	4

>>> isUndirectedGM gm1
False

>>> gm2
0	1	1
1	1	0
1	0	0

>>> isUndirected gm2
True

isDirectedG :: (Eq t, Num t) => Graph t -> Bool

The isDirectedG function returns whether a graph is directed or not. A graph is directed if it is not equal to its transpose. The function takes one argument, which is of type Graph.

Below are a few examples:

>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> isDirectedG g
True

isDirectedGM :: (Num a, Eq a) => GraphMatrix a -> Bool

The isDirectedGM function returns whether a graph (adjacency matrix form) is directed or not. A graph is directed if it is not equal to its transpose. The function takes one argument, which is of type GraphMatrix.

Below are a few examples:

>>> gm1
1	3
0	5

>>> isDirected gm1
True

>>> gm2
1	2	5
3	5	7
0	0	4

>>> isDirectedGM gm2
True

unionG :: (Num a, Ord a) => Graph a -> Graph a -> Graph a

The unionG function returns the union of two graphs. The function takes two arguments, both of the type Graph, and returns a Graph type.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> g'
Graph ({1,2,3},{(1,2,1),(1,4,5)})

>>> unionG g g'
Graph ({1,2,3,4},{(1,2,1),(1,2,5),(1,3,7),(1,4,5),(2,4,3)})

unionGM :: (Integral a, Integral a1) => GraphMatrix a -> GraphMatrix a1 -> GraphMatrix Int

The unionGM function returns the union of two graphs (in their adjancency matrix representation). The function takes two arguments, both of the type GraphMatrix, and returns a GraphMatrix type,

For example:

>>> gm
0	5	7	0
0	0	0	3
0	0	0	0
0	0	0	0

>>> gm'
0	1	0
0	0	0
0	0	0

>>> unionGM gm gm'
0	1	7	0
0	0	0	3
0	0	0	0
0	0	0	0

addVerticesG :: (Num a, Eq a) => Graph a -> Vertices a -> Graph a

The addVerticesG function is used to add vertices to a graph. The function takes two arguments, one of type Graph and the other of type Vertices. The function adds the vertices and returns a Graph.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,5),(1,3,7),(2,4,3)})

>>> addVerticesG g (Vertices [5])
Graph ({1,2,3,4,5},{(1,2,5),(1,3,7),(2,4,3)})

>>> addVerticesG g (Vertices [5..8])
Graph ({1,2,3,4,5,6,7,8},{(1,2,5),(1,3,7),(2,4,3)})

addVerticesGM :: Integral a => GraphMatrix a -> Vertices Int -> GraphMatrix Int

The addVerticesGM function is used to add vertices to a graph (adjacency matrix representation. The function takes two arguments, one of type GraphMatrix and the other of type Vertices. The function adds the vertices and returns a GraphMatrix.

For example:

>>> gm
0	5	7	0
0	0	0	3
0	0	0	0
0	0	0	0

>>> addVerticesGM gm (Vertices [5])
0	5	7	0	0
0	0	0	3	0
0	0	0	0	0
0	0	0	0	0
0	0	0	0	0

>>> addVerticesGM gm (Vertices [5,6])
0	5	7	0	0	0
0	0	0	3	0	0
0	0	0	0	0	0
0	0	0	0	0	0
0	0	0	0	0	0
0	0	0	0	0	0

getVerticesFromEdges :: Eq a => Edges a -> [a]

The getVerticesFromEdges function returns a list of vertices that form edges. The function takes one argument of type Edges and retuns a list.

For example:

>>> e
{(1,2,5),(1,3,7),(2,4,3)}

>>> getVerticesFromEdges e
[1,2,3,4]

addEdgesG :: (Num a, Eq a) => Graph a -> Edges a -> Graph a

The addEdgesG function adds edges to a graph. The new edges must consist of the vertices already existing in the graph. The function takes two arguments, the first of type Graph and the second of type Edges. After adding the edges, the new Graph is returned.

For example:

>>> g
Graph ({1,2,3},{(1,2),(1,3)})

>>> addEdgesG g (Edges [(2,3),(3,1)])
Graph ({1,2,3},{(1,2),(1,3),(2,3),(3,1)})

>>> addEdgesG g (Edges [(2,3),(3,1),(4,1)])
Graph *** Exception: Vertices in the edge(s) are not in the graph's set of vertices.

addEdgesGM :: Integral a => GraphMatrix a -> Edges Int -> GraphMatrix Int

The addEdgesGM function adds edges to a graph (adjacency matrix representation). The new edges must consist of the vertices already existing in the graph. The function takes two arguments, the first of type GraphMatrix and the second of type Edges. After adding the edges, the new GraphMatrix is returned.

For example:

>>> gm
0	5	7	0
0	0	0	3
0	0	0	0
0	0	0	0

>>> addEdgesGM gm (Edges [(1,4,5),(2,3,6),(4,3,1)])
0	5	7	5
0	0	6	3
0	0	0	0
0	0	1	0

areConnectedGM :: (Eq a, Floating a) => GraphMatrix a -> Vertices Int -> Vertices Int -> Bool

The areConnectedGM function is used to check if two vertices are directly (or even indirectly) connected. The function takes three arguments, a GraphMatrix and two Vertices.

For example:

>>> gm1
0.0	1.0	1.0	0.0
1.0	0.0	0.0	1.0
1.0	0.0	0.0	1.0
0.0	1.0	1.0	0.0

>>> areConnectedGM gm1 (Vertices [1]) (Vertices [3])
True

numPathsBetweenGM :: Floating a => GraphMatrix a -> Vertices Int -> Vertices Int -> a

The numPathsBetweenGM function is used to find the number of paths that exist between two given vertices in a graph. The function takes two arguments, the first is the GraphMatrix and the other two are Vertices.

For example:

>>> gm1
0.0	1.0	1.0	0.0
1.0	0.0	0.0	1.0
1.0	0.0	0.0	1.0
0.0	1.0	1.0	0.0

>>> numPathsBetweenGM gm1 (Vertices [1]) (Vertices [4])
8.0

adjacentNodesG :: (Num a, Eq a) => Graph a -> Vertices a -> Vertices a

The adjacentNodesG function returns all the adjacent vertices of a mentioned vertex. The function takes two arguments, the first is a Graph and the second is a vertex of type Vertices.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,1),(1,3,4),(2,1,5),(2,4,1),(3,1,2),(3,4,4),(4,2,1),(4,3,1)})

>>> adjacentNodes g (Vertices [1])
{2,3}

>>> adjacentNodes g (Vertices [2])
{1,4}

>>> adjacentNodes g (Vertices [3])
{1,4}

>>> adjacentNodes g (Vertices [4])
{2,3}

adjacentNodesGM :: Integral a => GraphMatrix a -> Vertices Int -> Vertices Int

The adjacentNodesGM function returns all the adjacent vertices of a mentioned vertex. The function takes two arguments, the first is a GraphMatrix and the second is a vertex of type Vertices.

For example:

>>> gm
0	1	4	0
5	0	0	1
2	0	0	4
0	1	1	0

>>> adjacentNodesGM gm (Vertices [1])
{2,3}

>>> adjacentNodesGM gm (Vertices [2])
{1,4}

>>> adjacentNodesGM gm (Vertices [3])
{1,4}

>>> adjacentNodesGM gm (Vertices [4])
{2,3}

inDegreeG :: (Num a, Eq a) => Graph a -> Vertices a -> Int

The inDegreeG function returns the in-degree of a vertex in a directed graph. The function takes two arguments, the first is a Graph and the second is a vertex of type Vertices

For example:

>>> g
Graph ({1,2,3,4},{(1,2,1),(1,3,4),(2,1,5),(2,4,1),(3,1,2),(3,4,4),(4,2,1),(4,3,1)})

>>> inDegreeG g (Vertices [1])
2

inDegreeGM :: Integral a => GraphMatrix a -> Vertices Int -> Int

The inDegreeGM function returns the in-degree of a vertex in a directed graph. The function takes two arguments, the first is a GraphMatrix and the second is a vertex of type Vertices

For example:

>>> gm
0	1	4	0
5	0	0	1
2	0	0	4
0	1	1	0

>>> inDegreeGM gm (Vertices [2])
2

outDegreeG :: (Eq a, Num a) => Graph a -> Vertices a -> Int

The outDegreeG function returns the out-degree of a vertex in a directed graph. The function takes two arguments, the first is a Graph and the second is a vertex of type Vertices.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,1),(1,3,4),(2,1,5),(2,4,1),(3,1,2),(3,4,4),(4,2,1),(4,3,1)})

>>> outDegree g (Vertices [4])
2

>>> outDegree g (Vertices [1])
2

outDegreeGM :: Integral a => GraphMatrix a -> Vertices Int -> Int

The outDegreeGM function returns the out-degree of a vertex in a directed graph. The function takes two arguments, the first is a GraphMatrix and the second is a vertex of type Vertices.

For example:

>>> gm
0	1	4	0
5	0	0	1
2	0	0	4
0	1	1	0

>>> outDegreeGM gm (Vertices [3])
2

degreeG :: (Eq a, Num a) => Graph a -> Vertices a -> Int

The degreeG function returns the degree of a vertex in a undirected graph. The function takes two arguments, the first is the undirected Graph and the second is a vertex of type Vertices.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,1),(1,3,4),(2,1,5),(2,4,1),(3,1,2),(3,4,4),(4,2,1),(4,3,1)})

>>> degreeG g (Vertices [4])
4

degreeGM :: Integral a => GraphMatrix a -> Vertices Int -> Int

The degreeGM function returns the degree of a vertex in a undirected graph (adjacency matrix representation. The function takes two arguments, the first is the undirected GraphMatrix and the second is a vertex of type Vertices.

For example:

>>> gm
0	1	4	0
5	0	0	1
2	0	0	4
0	1	1	0

>>> degreeGM gm (Vertices [2])
4

hasEulerCircuitG :: (Eq a, Num a) => Graph a -> Bool

The hasEulerCircuitG function checks if a specified graph conrtains a Euler Circuit.

For example:

>>> g
Graph ({1,2,3,4,5},{(1,2,1),(1,5,1),(2,1,1),(2,5,1),(3,4,1),(3,5,1),(4,3,1),(4,5,1),(5,1,1),(5,2,1),(5,3,1),(5,4,1)})

>>> hasEulerCircuitG g
True

hasEulerCircuitGM :: Integral a => GraphMatrix a -> Bool

The hasEulerCircuitGM function checks if a specified graph (adjacency matrix representation) conrtains a Euler Circuit.

For example:

>>> gm
0	1	0	0	1
1	0	0	0	1
0	0	0	1	1
0	0	1	0	1
1	1	1	1	0

>>> hasEulerCircuitGM gm
True

hasEulerPathG :: (Eq a, Num a) => Graph a -> Bool

The hasEulerPathG function checks if a graph contains a Euler Path.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,1),(1,4,1),(2,1,1),(2,3,1),(2,4,1),(3,2,1),(3,4,1),(4,1,1),(4,2,1),(4,3,1)})

>>> hasEulerPathG g
True

hasEulerPathGM :: Integral a => GraphMatrix a -> Bool

The hasEulerPathGM function checks if a graph (adjacency matrix representation) contains a Euler Path.

For example:

>>> gm
0	1	0	1
1	0	1	1
0	1	0	1
1	1	1	0

>>> hasEulerPathGM gm
True

hasHamiltonianCircuitG :: (Eq a, Num a) => Graph a -> Bool

The hasHamiltonianCircuitG function checks if a graph contains a Hamiltonian Circuit.

For example:

>>> g2
Graph ({1,2,3,4,5},{(1,2,1),(1,3,1),(1,5,1),(2,1,1),(2,3,1),(2,5,1),(3,1,1),(3,2,1),(3,4,1),(3,5,1),(4,3,1),(4,5,1),(5,1,1),(5,2,1),(5,3,1),(5,4,1)})

>>> hasHamiltonianCircuitG g2
True

hasHamiltonianCircuitGM :: Integral a => GraphMatrix a -> Bool

The hasHamiltonianCircuitGM function checks if a graph (adjacency matrix representation) contains a Hamiltonian Circuit.

For example:

>>> gm
0	1	1	0	1
1	0	1	0	1
1	1	0	1	1
0	0	1	0	1
1	1	1	1	0

>>> hasHamiltonianCircuitGM gm
True

countOddDegreeV :: (Eq a1, Num a, Num a1) => Graph a1 -> a

The countOddDegreeV function returns the number of vertices in a graph that have an odd degree.

For example:

>>> g
Graph ({1,2,3,4},{(1,2,1),(1,4,1),(2,1,1),(2,3,1),(2,4,1),(3,2,1),(3,4,1),(4,1,1),(4,2,1),(4,3,1)})

>>> countOddDegreeV g
0

>>> g1
Graph ({1,2,3,4},{(1,2,4),(1,3,1),(1,4,5),(3,2,6)})

>>> countOddDegreeV g1
2

countEvenDegreeV :: (Eq a1, Num a, Num a1) => Graph a1 -> a

The countEvenDegreeV function returns the number of vertices in a graph that have an even degree.

For example:

>>> g1
Graph ({1,2,3,4},{(1,2,4),(1,3,1),(1,4,5),(3,2,6)})

>>> countEvenDegreeV g1
2

hasEulerPathNotCircuitG :: (Eq a, Num a) => Graph a -> Bool

The hasEulerPathNotCircuitG function checks if a graph conatains a Euler Path but not a Euler circuit.

For example:

>>> g1
Graph ({1,2,3,4},{(1,2,4),(1,3,1),(1,4,5),(3,2,6)})

>>> hasEulerPathNotCircuitG g1
True

hasEulerPathNotCircuitGM :: Integral a => GraphMatrix a -> Bool

The hasEulerPathNotCircuitGM function checks if a graph (adjacency matrix representation) conatains a Euler Path but not a Euler circuit.

For example:

>>> gm
0	4	1	5
0	0	0	0
0	6	0	0
0	0	0	0

>>> hasEulerPathNotCircuitGM gm
True

isSubgraphG :: (Num a, Ord a) => Graph a -> Graph a -> Bool

The isSubgraphG function checks if a graph is a sub-graph if another specified graph.

For example:

>>> g1
Graph ({1,2,3,4},{(1,2,4),(1,3,1),(1,4,5),(3,2,6)})

>>> g2
Graph ({1,3},{(1,3,1)})

>>> isSubgraphG g2 g1
True

>>> isSubgraphG g1 g2
False

isSubgraphGM :: (Integral a, Integral a1) => GraphMatrix a -> GraphMatrix a1 -> Bool

The isSubgraphGM function checks if a graph is a sub-graph if another specified graph (both graphs as GraphMatrix - adjacency matrix representation).

For example:

>>> grm1
0	4	1	5
0	0	0	0
0	6	0	0
0	0	0	0

>>> grm2
0	4
0	0

>>> isSubgraphGM grm2 grm1
True

>>> isSubgraphGM grm1 grm2
False