Matrix

Description

Functionality for: * Matrix addition * Matrix subtraction * Matrix multiplication * Multiplication with a scalar * Matrix division * Matrix transposition * Matrix inversion * Finding number of rows and columns * Matrix equality * Extractig rows and columns * Power of matrices * Trace of a matrix * Checking for type of matrix (symmetric, skew-symmetric, row, column, square, zero, one, unit/identity, orthogonal, involutive, invertible) * Determinant * Calculating determinant of a matrix * Finding minors and cofactors

Documentation

newtype Matrix a

Matrix is the data type used for representing a matrix as a two-dimensional list.

For example:

>>> let m = Matrix [[1,2,3],[4,5,6],[7,8,9]]
>>> m
1   2   3
4   5   6
7   8   9

Constructors

Matrix [[a]]

mAdd :: Num a => Matrix a -> Matrix a -> Matrix a

The mAdd function adds two matrices, each of type Matrix, and returns their sum as a Matrix.

For example:

>>> mAdd (Matrix [[1,2,3],[4,5,6]]) (Matrix [[1,1,1],[0.2,0.5,0.5]])
2.0     3.0     4.0
4.2     5.5     6.5

mAddL :: Num a => [Matrix a] -> Matrix a

The mAddL function can be used to add a list of matrices, each of type Matrix.

For example:

>>> m1
1   0
0   1

>>> m4
1   6
2   3

>>> mAddL [m1,m4]
2   6
2   4

(|+|) :: Num a => Matrix a -> Matrix a -> Matrix a

The |+| operator can also be used for adding two matrices.

For example:

>>> (Matrix [[1,2,3]]) |+| (Matrix [[1,1,1]])
2   3   4

mSub :: Num a => Matrix a -> Matrix a -> Matrix a

The mSub function is used to subtract one matrix from another, both of type Matrix and return their difference as a Matrix.

For example:

mSub (Matrix [[1,2,3],[4,5,0]]) (Matrix [[1,1,1],[0.2,0.5,0.5]]) >>> 0.0 1.0 2.0 >>> 3.8 4.5 -0.5

(|-|) :: Num a => Matrix a -> Matrix a -> Matrix a

The |-| can also be used to find the difference between two matrices.

For example:

(Matrix [[1,2,3],[4,5,6]]) |-| (Matrix [[1,1,1],[0,0,0]]) >>> 0 1 2 >>> 4 5 6

mTranspose :: Matrix a -> Matrix a

The mTranspose function returns the transpose of a matrix.

For example:

>>> mTranspose (Matrix [[1,0,1],[0,1,2]])
1 0
0 1
1 2

>>> mTranspose (Matrix [[1,0],[2,1],[3,2],[4,3]])
1 2 3 4
0 1 2 3

mScalarMult :: Num a => a -> Matrix a -> Matrix a

The mScalarMult function is used to multiply a scalar value and a matrix .

For example:

>>> mScalarMult (-3/2) (Matrix [[1,0,0],[0,1,0],[0,0,1]])
-1.5 -0.0 -0.0
-0.0 -1.5 -0.0
-0.0 -0.0 -1.5

>>> mScalarMult 3 (Matrix [[1,0,0],[0,1,0],[0,0,1]])
3 0 0
0 3 0
0 0 3

(|*|) :: Num a => a -> Matrix a -> Matrix a

The |*| operator can also be used to multiply a scalar and a matrix.

For example:

>>> (-3/2) |*| (Matrix [[1,0,0],[0,1,0],[0,0,1]])
-1.5 -0.0 -0.0
-0.0 -1.5 -0.0
-0.0 -0.0 -1.5

mMult :: Num a => Matrix a -> Matrix a -> Matrix a

The mMult function multiplies two matrices and returns the product.

For example:

>>> mMult (Matrix [[1,2,3],[4,5,6]]) (Matrix [[4.5,7,8],[10,(-10),6]])
24.5    -13.0       20.0
68.0    -22.0       62.0

mMultL :: Num a => [Matrix a] -> Matrix a

The mMultL function multiplies a list of matrices and returns their product.

For example:

>>> m1
1   0
0   1

>>> m4
1   6
2   3

>>> mMultL [m1,m1,m4]
1   6
2   3

(|><|) :: Num a => Matrix a -> Matrix a -> Matrix a

The |><| function can also be used to multiply two matrices.

For example:

>>> (Matrix [[1,0],[0,1]]) |><| (Matrix [[4.5,8],[(-10),6]])
4.5 8.0
-10.0   6.0

numRows :: Num a => Matrix a -> Int

The numRows function returns the number of rows in a specified matrix..

For example:

>>> numRows (Matrix [[1],[2],[3],[4]])
4

>>> numRows (Matrix [])
0

>>> numRows (Matrix [[]])
1

>>> numRows (Matrix [[1..10],[2,4..20]])
2

numCols :: Num a => Matrix a -> Int

The numCols function returns the number of columns in a specified matrix.

For example:

>>> numCols (Matrix [[1],[2],[3],[4]])
1

>>> numCols (Matrix [])
0

>>> numCols (Matrix [[]])
1

>>> numCols (Matrix [[1..10],[2,4..20]])
10

mat2list :: Matrix a -> [[a]]

The mat2list function converts a Matrix into a two-dimensional list.

For example:

>>> mat2list (Matrix [[1..10],[2,4..20]])
[[1,2,3,4,5,6,7,8,9,10],[2,4,6,8,10,12,14,16,18,20]]

>>> mat2list (Matrix [[1],[2],[3],[4]])
[[1],[2],[3],[4]]

determinant :: Num a => Matrix a -> a

The determinant function returns the determinant of a matrix.

For example:

>>> determinant (Matrix [[1,0],[0,1]])
1.0

>>> determinant (Matrix [[1,1],[1,(-1)]])
-2

>>> determinant (Matrix [[2,4,5],[8,(-1.5),6.8]])
-35.0

inverse :: (Eq a, Fractional a) => Matrix a -> Matrix a

The inverse function returns the inverse of a specified matrix.

For example:

>>> let m5 = Matrix [[4,3],[3,2]]
>>> inverse m5
-2.0    3.0
3.0     -4.0

>>> let m6 = Matrix [[1,2,3],[0,4,5],[1,0,6]]
>>> inverse m6
1.0909090909090908      -0.5454545454545454     -9.090909090909091e-2
0.2272727272727273      0.13636363636363635     -0.2272727272727273
-0.18181818181818182    9.090909090909091e-2    0.18181818181818182

>>> inverse (Matrix [[1,0],[0,1]])
1.0     -0.0
-0.0    1.0

>>> inverse (Matrix [[4.6,3.5],[3.1,2]])
-1.212121212121212  2.121212121212121
1.8787878787878785  -2.787878787878787

mDiv :: (Eq a, Fractional a) => Matrix a -> Matrix a -> Matrix a

The mDiv function is used to divide a matrix by another.

For example:

>>> mDiv (Matrix [[4.5,6.7],[8.9,10.0]]) (Matrix [[1.0,0.0],[0.0,1.0]])
4.5 6.7
8.9 10.0

(|/|) :: (Eq a, Fractional a) => Matrix a -> Matrix a -> Matrix a

The |/| operator can also be used to divide two matrices.

For example:

>>> Matrix [[1.0,0.0],[0.0,1.0]]) |/| (Matrix [[4.5,6.7],[8.9,10.0]])
-0.6835269993164729 0.45796308954203685
0.608339029391661 -0.3075871496924128

extractRow :: Matrix a -> Int -> [a]

The extractRow function returns a particular row of the matrix passed as argument.

For example:

>>> extractRow (Matrix [[1..5]]) 0
[1,2,3,4,5]

>>> extractRow (Matrix [[1..5],[6..10],[11..15]]) 2
[11,12,13,14,15]

extractCol :: Matrix a -> Int -> [a]

The extractCol function returns a particular column of the matrix passed as argument.

For example:

>>> extractCol (Matrix [[1..5],[6..10],[11..15]]) 2
[3,8,13]

>>> extractCol (Matrix [[1..5]]) 0
[1]

extractRowRange :: Matrix a -> Int -> Int -> Matrix a

The extractRowRange function returns a range of rows from the passed matrix.

For example:

>>> extractRowRange (Matrix [[1,2],[3,4],[5,6],[7.8,(-9)],[10,11.5]]) 2 4
5.0     6.0
7.8     -9.0
10.0    11.5

extractColRange :: Matrix a -> Int -> Int -> Matrix a

The extractColRange function returns a range of columns from the passed matrix.

For example:

>>> extractColRange (Matrix [[1,2],[3,4],[5,6],[7.8,(-9)],[10,11.5]]) 0 1
1.0 3.0 5.0 7.8 10.0
2.0 4.0 6.0 -9.0 11.5

mPower :: (Eq a1, Fractional a1, Num a, Ord a) => Matrix a1 -> a -> Matrix a1

The mPower function returns the power of a matrix.

For example:

>>> m5
4.6 3.5
3.1 2.0
>>> mPower m5 10
1.5349250119379473e8    1.1096944577877794e8
9.828722340406047e7     7.105805575813112e7

>>> mPower (unit 3) 3
1.0 0.0 0.0
0.0 1.0 0.0
0.0 0.0 1.0

>>> mPower (unit 3) (-3)
1.0 0.0 0.0
0.0 1.0 0.0
0.0 0.0 1.0

trace :: Num a => Matrix a -> a

The trace function returns the trace of a matrix.

For example:

>>> trace (Matrix [[1,1,1],[1,2,3],[4,(-6),7.8]])
10.8

>>> trace (Matrix [[1,0,0],[0,1,0],[0,0,1]])
3

isInvertible :: (Eq a, Num a) => Matrix a -> Bool

The isInvertible function checks if a matrix can be inverted or not.

For example:

>>> isInvertible (Matrix [[1,1],[1,(-1)]])
True

>>> isInvertible (Matrix [[1,1],[(-1),1]])
True

>>> isInvertible (Matrix [[1,1],[1,1]])
False

>>> isInvertible (Matrix [[1,0],[0,1]])
True

isSymmetric :: Eq a => Matrix a -> Bool

The isSymmetric function checks if a matrix is symmetric.

For example:

>>> isSymmetric (Matrix [[1,2,3],[2,1,3],[3,3,4]])
True

>>> isSymmetric (Matrix [[1,0],[0,1]])
True

>>> isSymmetric (Matrix [[1,1],[(-1),1]])
False

isSkewSymmetric :: (Eq a, Num a) => Matrix a -> Bool

The isSkewSymmetric function checks if a matrix is anti/skew-symmetric.

For example:

>>> isSkewSymmetric (Matrix [[0,0],[0,0]])
True

>>> isSkewSymmetric (Matrix [[1,1],[(-1),1.5]])
False

isRow :: Num a => Matrix a -> Bool

The isRow function checks if a matrix is a row matrix.

For example:

>>> isRow (Matrix [[1..5]])
True

>>> isRow (Matrix [[1..5],[2..7]])
False

isColumn :: Num a => Matrix a -> Bool

The isColumn function checks if a matrix is a column matrix.

For example:

>>> isColumn (Matrix [[1..4],[2..6]])
False

>>> isColumn (Matrix [[1],[2],[3]])
True

isSquare :: Num a => Matrix a -> Bool

The isSquare matrix checks if a matrix is a square matrix.

For example:

>>> isSquare (Matrix [[1,0,0],[1,2,3]])
False

>>> isSquare (Matrix [[1,0],[0,1]])
True

isOrthogonal :: (Eq a, Fractional a) => Matrix a -> Bool

The isOrthogonal function checks if a matrix is orthogonal.

For example:

>>> isOrthogonal (Matrix [[1,0],[0,1]])
True

>>> isOrthogonal (Matrix [[1,1],[1.2,(-1.5)]])
False

isInvolutive :: (Eq a, Fractional a) => Matrix a -> Bool

The isInvolutive function checks if a matrix is involutive.

For example:

>>> isInvolutive (Matrix [[1,0],[0,1]])
True

>>> isInvolutive (Matrix [[1,2],[2,1]])
False

isZeroOne :: (Eq a, Num a) => Matrix a -> Bool

The isZeroOne function returns true if a matrix is a zero or one matrix.

For example:

>>> isZeroOne (Matrix [[1,1],[1,4]])
False

>>> isZeroOne (Matrix [[1,1],[1,0]])
True

isZero :: (Eq a, Num a) => Matrix a -> Bool

The isZero function returns true if a matrix is a zero matrix (all elements evaluate to 0).

For example:

>>> isZero (Matrix [[0,0,0]])
True

>>> isZero (Matrix [[0,1],[1,2]])
False

isOne :: (Eq a, Num a) => Matrix a -> Bool

The isOne function returns true if a matrix is a one matrix (all elements evaluate to 1). Usage: isOne (Matrix [[1,1],[1,1]]) >>> True

isOne (Matrix [[1,1],[1,0]]) >>> False

isUnit :: (Eq a, Fractional a) => Matrix a -> Bool

The isUnit function returns true is a matrix is a unit matrix.

For example:

>>> isUnit (Matrix [[1,0],[0,1]])
True

>>> isUnit (Matrix [[3,0],[0,3]])
False

>>> isUnit (Matrix [[1,0,0],[0,1,0],[0,0,1]])
True

>>> isUnit (Matrix [[]])
False

>>> isUnit (Matrix [[1]])
True

mMap :: (a1 -> a) -> Matrix a1 -> Matrix a

The mMap function maps a function to a matrix, i.e. applies a function to all elements of a matrix.

For example:

>>> m
1.0 0.0 2.0
2.0 5.0 6.5

>>> mMap (*3) m
3.0     0.0     6.0
6.0     15.0    19.5

zero :: Num a => Int -> Matrix a

The zero function returns a zero matrix of order 'n x n', where n is the argument passed.

For example:

>>> zero 2
0   0
0   0

>>> zero 3
0   0   0
0   0   0
0   0   0

zero' :: Num a => Int -> Int -> Matrix a

The zero' function returns a zero matrix of order 'm x n', where m and n are the arguments passed.

For example:

>>> zero 2 3
0   0   0
0   0   0

>>> zero 3 3
0   0   0
0   0   0
0   0   0

one :: Num a => Int -> Matrix a

The one function returns a one matrix of order 'n x n', where n is the argument passed.

For example:

>>> one 2
1   1
1   1

>>> one 3
1   1   1
1   1   1
1   1   1

one' :: Num a => Int -> Int -> Matrix a

The one' function returns a one matrix of order 'm x n', where m and n are the arguments passed.

For example:

>>> one 2 2
1   1
1   1

>>> one 3 2
1   1
1   1
1   1

unit :: Num a => Int -> Matrix a

The unit function returns a unit matrix of order 'n x n', where n is the argument passed.

For example:

>>> unit 2
1   0
0   1

>>> unit 3
1   0   0
0   1   0
0   0   1