Modular

Description

Functionality for: * Modular addition * Modular subtraction * Modular multiplication * Modular exponentiation * Checking congruence * Solving linear congruences

Documentation

modAdd :: Integral a => a -> a -> a -> a

The modAdd function performs addition using modular arithmetic.

modAdd a b m = (a + b) mod m

(a + b) mod m = (a mod m) + (b mod m)

For example:

>>> modAdd 17 44 11
6

>>> modAdd 126832 1832 11
8

modSub :: Integral a => a -> a -> a -> a

The modSub function performs subtraction using modular arithmetic.

modSub a b m = (a - b) mod m

(a - b) mod m = (a mod m) - (b mod m)

For example:

>>> modSub 117 14 11
4

>>> modSub 114787 23934 3874
1751

modMult :: Integral a => a -> a -> a -> a

The modMult function performs multiplication using modular arithmetic.

modMult a b m = (a * b) mod m

(a * b) mod m = ((a mod m) * (b mod m)) mod m

For example:

>>> modMult 117 14 11
10

>>> modMult 114787 23934 3874
2974

modExp :: (Integral a, Integral a1) => a -> a1 -> a -> a

The modExp function is for the Modular Exponentiation operation (a ^ b mod m)

For example:

>>> modExp 12 5 6
0

>>> modExp 112 34 546
532

>>> modExp 515 5151 1563
1004

isCongruent :: Integral a => a -> a -> a -> Bool

The isCongruent function checks for modular congruency. For a = b (mod m) to be congruent, a mod m = b.

For example:

>>> isCongruent 132 2 130
True

>>> isCongruent 13493 238 234
False

findCongruentPair :: Integral t => t -> t -> t -> t -> [t]

The findCongruentPair function uses the linear congruence formula a * x = b (mod m) or (a * x) mod m = b to find x, when a, b and m are given.

For example:

>>> findCongruentPair 5 6 199 100
[41]

>>> isCongruent (5 * 41) 6 199
True

>>> findCongruentPair 5 6 199 500
[41,240,439]

findCongruentPair1 :: Integral t => t -> t -> t -> t -> [t]

The findCongruentPair1 function uses the linear congruence formula a + b = b (mod m) or (a + x) mod m = b to find x, when a, b and m are given.

For example:

>>> findCongruentPair1 10 4 5 100
[4,9,14,19,24,29,34,39,44,49,54,59,64,69,74,79,84,89,94,99]

>>> findCongruentPair1 113 20 40 100
[27,67]

>>> isCongruent (113 + 27) 20 40
True

>>> isCongruent (113 + 67) 20 40
True