Primes

Description

Functionality for: * Generating prime numbers upto a limit (Sieve of Eratosthenes) * Generate first n primes * Generate the immediately next prime number greater than or equal to given number * Primality testing using Trial division and Miller-Rabin test * Prime factorization

Documentation

primesTo :: Integer -> [Integer]

The primesTo function generates a list of prime numbers upto a specified limit by using Sieve of Eratosthenes

For example:

>>> primesTo 20
[2,3,5,7,11,13,17,19]

>>> primesTo 76
[2,3,5,7,11,13,17,19,23,29,31,37,41,43,47,53,59,61,67,71,73]

primesBetween :: Integer -> Integer -> [Integer]

The primesBetween function returns all prime numbers between the mention lower and upper limits (upper limit inclusive)

For example:

>>> primesBetween 0 10
[2,3,5,7]

>>> primesBetween 10 25
[11,13,17,19,23]

>>> primesBetween 5 2
[2,3,5]

nPrimes :: Integral a => Int -> [a]

The nPrimes function returns a list of first n primes, where n is the argument passed.

For example:

>>> nPrimes 10
[2,3,5,7,11,13,17,19,23,29]

>>> nPrimes 42
[2,3,5,7,11,13,17,19,23,29,31,37,41,43,47,53,59,61,67,71,73,79,83,89,97,101,103,107,109,113,127,131,137,139,149,151,157,163,167,173,179,181]

primesTo100 :: [Integer]

The primesTo100 function returns a list of all prime numbers less than 100.

primesTo10000 :: [Integer]

The primesTo100 function returns a list of all prime numbers less than 10000.

isTrialDivisionPrime :: Integer -> Bool

The isTrialDivisionPrime function is checks the primality of a number using trial division.

For example:

>>> isTrialDivisionPrime 83498594
False

>>> isTrialDivisionPrime 83498617
True

>>> isTrialDivisionPrime 83498594823549879237544
False

isMillerRabinPrime :: Integer -> Bool

The isMillerRabinPrime function checks if number is prime using Miller-Rabin primality test

For example:

>>> isMillerRabinPrime 34658273497773875
False

>>> isMillerRabinPrime 34658273497773875327483287487877599
True

isPrime :: Integer -> Bool

The isPrime function checks if a number is prime.

The function uses the isTrialDivisionPrime or isMillerRabinPrime functions to test the primality.

For example:

>>> isPrime (5^100)
False

>>> isPrime 3973793488238
False

isPrime 199 >>> True

nextPrime :: Integer -> Integer

The nextPrime function returns the next prime greater than or equal to the passed argument.

For example:

>>> nextPrime 10240
10243

>>> nextPrime 3894345
3894347

primeFactors :: Integer -> [Integer]

The primeFactors function returns all the prime factors of a number.

The function includes repetitions of factors, as shown in the examples below. To get only the unique factors, use the uniquePrimeFactors function.

For example:

>>> primeFactors 23984
[2,2,2,2,1499]

>>> primeFactors 273219943
[19,89,161573]

uniquePrimeFactors :: Integer -> [Integer]

The uniquePrimeFactors function returns a single instance of all the prime factors of a number.

For example:

>>> uniquePrimeFactors 10000000
[2,5]
>>> uniquePrimeFactors 123456789
[3,3607,3803]

areCoprime :: Integer -> Integer -> Bool

The areCoprime function checks if two numbers are coprime relatively prime mutually prime.

For example:

>>> areCoprime 2 3
True

>>> areCoprime 100 200
False