Relation

Description

Functionality for * Generating element set * Obtaining list of first element values * Obtaining list of second element values * Obtaining list of first element values for a specified element as a second element * Obtaining list of second element values for a specified element as a first element * Checking for reflexivity, symmetricity, anti-symmetricity, transitivity * Union, intersection and difference of two relations * Relation composition * Power of relations * Reflexive, Symmetric and Transitive closures

Documentation

newtype Relation a

The Relation data type is used for represnting relations (discrete mathematics).

For example:

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> r2
{(1,1),(2,2),(3,3)}

Constructors

Relation [(a, a)]

relation2list :: Relation t -> [(t, t)]

The relation2list function converts a Relation to a list representation.

For example:

>>> r2
{(1,1),(2,2),(3,3)}

>>> relation2list r2
[(1,1),(2,2),(3,3)]

list2relation :: [(a, a)] -> Relation a

The list2relation function converts a list to a relation.

For example:

>>> let l = [(1,2),(2,3),(1,3)]
>>> let r = list2relation l
>>> r
{(1,2),(2,3),(1,3)}

inverse :: Relation a -> Relation a

The inverse function returns the inverse Relation of a specified Relation.

For example:

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> inverse r1
{(1,1),(2,1),(3,1),(1,2),(2,2),(3,2),(1,3),(2,3),(3,3)}

>>> r2
{(1,1),(2,2),(3,3)}

>>> inverse r2
{(1,1),(2,2),(3,3)}

getDomain :: Eq a => Relation a -> [a]

The getDomain function returns the list of all "a" where (a,b) <- Relation.

For example:

>>> getDomain (Relation [(1,2),(3,4),(2,5)])
[1,3,2]

>>> getDomain (Relation [])
[]

getRange :: Eq a => Relation a -> [a]

The getRange function returns the list of all "b" where (a,b) <- Relation.

For example:

>>> getRange (Relation [(1,2),(3,4),(2,5)])
[2,4,5]

>>> getRange (Relation [])
[]

elements :: Eq a => Relation a -> [a]

The elements function returns a list of all elements in a Relation.

For example:

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> elements r1
[1,2,3]

returnDomainElems :: Eq a => Relation a -> a -> [a]

The returnFirstElems function returns alist of all "a" where (a,b) <- Relation and "b" is specified

For example:

>>> returnFirstElems (Relation [(1,2),(1,3),(2,3),(3,3),(3,4)]) 1
[]

>>> returnFirstElems (Relation [(1,2),(1,3),(2,3),(3,3),(3,4)]) 4
[3]

>>> returnFirstElems (Relation [(1,2),(1,3),(2,3),(3,3),(3,4)]) 3
[1,2,3]

returnRangeElems :: Eq a => Relation a -> a -> [a]

The returnSecondElems function returns list of all b where (a,b) <- Relation and a is specified/

For example:

>>> returnSecondElems (Relation [(1,2),(1,3),(2,3),(3,3),(3,4)]) 3
[3,4]

>>> returnSecondElems (Relation [(1,2),(1,3),(2,5)]) 1
[2,3]

isReflexive :: Eq t => Relation t -> Bool

The isReflexive function checks if a Relation is reflexive or not.

For example:

>>> isReflexive (Relation [(1,1),(1,2),(2,2),(2,3)])
False

>>> isReflexive (Relation [(1,1),(1,2),(2,2)])
True

isIrreflexive :: Eq t => Relation t -> Bool

The isIrreflexive function checks if a Relation is irreflexive or not.

For example:

>>> isIrreflexive (Relation [(1,1),(1,2),(2,2),(2,3)])
True

>>> isIrreflexive (Relation [(1,1),(1,2),(2,2)])
False

isSymmetric :: Eq a => Relation a -> Bool

The isSymmetric function checks if a Relation is symmetric or not.

For example:

>>> isSymmetric (Relation [(1,1),(1,2),(2,2)])
False

>>> isSymmetric (Relation [(1,1),(1,2),(2,2),(2,1)])
True

isAsymmetric :: Eq t => Relation t -> Bool

The isAsymmetric function checks if a Relation is asymmetric or not.

For example:

>>> isAntiSymmetric (Relation [(1,2),(2,1)])
False

>>> isAntiSymmetric (Relation [(1,2),(1,3)])
True

isAntiSymmetric :: Eq a => Relation a -> Bool

The isAntiSymmetric function checks if a Relation is anti-symmetric or not.

For example:

>>> r2
{(1,1),(2,2),(3,3)}

>>> isAntiSymmetric r2
True

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> isAntiSymmetric r1
False

isTransitive :: Eq a => Relation a -> Bool

The isTransitive function checks if a Relation is transitive or not.

For example:

>>> isTransitive (Relation [(1,1),(1,2),(2,1)])
False

>>> isTransitive (Relation [(1,1),(1,2),(2,1),(2,2)])
True

>>> isTransitive (Relation [(1,1),(2,2)])
True

rUnion :: Ord a => Relation a -> Relation a -> Relation a

The rUnion function returns the union of two relations.

For example:

>>> rUnion (Relation [(1,1),(1,2)]) (Relation [(2,3),(2,2)])
{(1,1),(1,2),(2,2),(2,3)}

>>> rUnion (Relation [(1,1),(1,2)]) (Relation [(1,1)])
{(1,1),(1,2)}

rUnionL :: (Ord t1, Foldable t) => t (Relation t1) -> Relation t1

The rUnionL function returns the union of a list of Relation.

For example:

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> r2
{(1,1),(2,2),(3,3)}

>>> rUnionL [r1,r2]
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

rIntersection :: Ord a => Relation a -> Relation a -> Relation a

The rIntersection function returns the intersection of two relations.

For example:

>>> rIntersection (Relation [(1,1),(1,2)]) (Relation [(1,1)])
{(1,1)}

>>> rIntersection (Relation [(1,1),(1,2)]) (Relation [(2,3),(2,2)])
{}

rIntersectionL :: (Ord a, Foldable t) => t (Relation a) -> Relation a

The rIntersectionL function returns the intersection of a list of Relation.

For example:

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> r2
{(1,1),(2,2),(3,3)}

>>> rIntersectionL [r1,r2]
{(1,1),(2,2),(3,3)}

rDifference :: Ord a => Relation a -> Relation a -> Relation a

The rDifference function returns the difference of two relations.

For example:

>>> rDifference (Relation [(1,1),(1,2)]) (Relation [(1,1)])
{(1,2)}

>>> rDifference (Relation [(1,1),(1,2)]) (Relation [(2,3),(2,2)])
{(1,1),(1,2)}

rComposite :: Eq a => Relation a -> Relation a -> Relation a

The rComposite function returns the composite / concatenation of two relations.

For example:

>>> rComposite (Relation [(1,1),(1,2)]) (Relation [(2,3),(2,2)])
{(1,2),(1,3)}

>>> rComposite (Relation [(1,1),(1,2)]) (Relation [(1,1)])
{(1,1)}

rPower :: (Eq a, Eq a1, Integral a1) => Relation a -> a1 -> Relation a

The rPower function returns the power of a Relation.

For example:

>>> let r4 = Relation [(1,2), (2,3), (2,4), (3,3)]
>>> r4
{(1,2),(2,3),(2,4),(3,3)}

>>> rPower r4 2
{(1,3),(1,4),(2,3),(3,3)}

>>> rPower r4 (-2)
{(3,1),(3,2),(3,3),(4,1)}
--| pow < 0 = rPower (Relation [(b, a) | (a, b) <- r]) (-pow)

reflClosure :: Ord a => Relation a -> Relation a

The reflClosure function returns the Reflecive Closure of a Relation.

For example:

>>> reflClosure (Relation [(1,1),(1,2),(4,5)])
{(1,1),(1,2),(2,2),(4,4),(4,5),(5,5)}

>>> reflClosure (Relation [(1,1),(1,3)])
{(1,1),(1,3),(3,3)}

symmClosure :: Ord a => Relation a -> Relation a

The symmClosure function returns the Symmetric Closure of a Relation.

For example:

>>> symmClosure (Relation [(1,1),(1,2),(4,5)])
{(1,1),(1,2),(2,1),(4,5),(5,4)}

>>> symmClosure (Relation [(1,1),(1,3)])
{(1,1),(1,3),(3,1)}

tranClosure :: Ord a => Relation a -> Relation a

The tranClosure function returns the Transitive Closure of a Relation.

For example:

>>> tranClosure (Relation [(1,1),(1,2),(2,1)])
{(1,1),(1,2),(2,1),(2,2)}

>>> tranClosure (Relation [(1,1),(1,2),(1,3),(2,2),(3,1),(3,2)])
{(1,1),(1,2),(1,3),(2,2),(3,1),(3,2),(3,3)}

isEquivalent :: Eq a => Relation a -> Bool

The isEquivalent function checks if a Relation is equivalent (reflexive, symmetric and transitive).

For example:

>>> r2
{(1,1),(2,2),(3,3)}

>>> isEquivalent r2
True

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> isEquivalent r1
True

>>> isEquivalent (Relation [(1,2), (2,3)])
False

isWeakPartialOrder :: Eq a => Relation a -> Bool

The isWeakPartialOrder function checks if a Relation is a weak partial order (reflexive, anti-symmetric and transitive).

For example:

>>> r1
{(1,1),(1,2),(1,3),(2,1),(2,2),(2,3),(3,1),(3,2),(3,3)}

>>> isWeakPartialOrder r1
False

>>> r2
{(1,1),(2,2),(3,3)}

>>> isWeakPartialOrder r2
True

isWeakTotalOrder :: Eq a => Relation a -> Bool

The isWeakTotalOrder function checks if a Relation is a Weak Total Order, i.e. it is a Weak Partial Order and for all "a" and "b" in Relation "r", (a,b) or (b,a) are elements of r.

isStrictPartialOrder :: Eq a => Relation a -> Bool

The isStrictPartialOrder function checks if a Relation is a Strict Partial Order, i.e. it is irreflexive, asymmetric and transitive.

isStrictTotalOrder :: Eq a => Relation a -> Bool

The isStrictTotalOrder function checks if a Relation is a Strict Total Order, i.e. it is a Strict Partial Order, and for all "a" and "b" in Relation "r", either (a,b) or (b,a) are elements of r or a == b.