Tree

Description

Functionality for * Binary Tree data structure * Inorder, Preorder and Postorder tree traversals * Inserting nodes in an element * Finding if an element is in a tree * Calculating Reflection of a tree (mirroring) * Finding Height of a tree * Finding depth of a node * Determining Size of a tree (number of nodes) * Checking if a tree is balanced

Documentation

data BinTree a

The BinTree datatype is used for representing binary trees.

Constructors

Leaf

Node a (BinTree a) (BinTree a)

inorder :: BinTree a -> [a]

The inorder function returns the nodes of a BinTree after inorder traversal.

For example:

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))

>>> inorder tree2
[1,2,3,4,5,6,7,8]

preorder :: BinTree a -> [a]

The preorder function returns the nodes of a BinTree after preorder traversal.

For example:

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))

>>> preorder tree2
[4,2,1,3,7,5,6,8]

postorder :: BinTree a -> [a]

The postorder function returns the nodes of a BinTree after postorder traversal.

For example:

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))

>>> postorder tree2
[1,3,2,6,5,8,7,4]

singleton :: a -> BinTree a

The singleton function returns a single childless node.

For example:

>>> singleton 10
Node 10 Leaf Leaf

>>> singleton 'a'
Node 'a' Leaf Leaf

>>> singleton 'A'
Node 'A' Leaf Leaf

treeInsert :: Ord a => a -> BinTree a -> BinTree a

The treeInsert function inserts a node into a BinTree.

The function takes two arguments - the first is the value to be inserted and the second is the BinTree into which the value is to be inserted.

For example:

>>> treeInsert 10 (Leaf)
Node 10 Leaf Leaf

>>> treeInsert 10 tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf (Node 10 Leaf Leaf)))

treeElem :: Ord a => a -> BinTree a -> Bool

The treeElem function checks if an elements exists in a BinTree by using binary search.

For example:

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))

>>> treeElem 5 tree2
True

>>> treeElem 10 tree2
False

reflect :: BinTree a -> BinTree a

The reflect function returns the mirror reflect of a BinTree.

For example:

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))

>>> inorder tree2
[1,2,3,4,5,6,7,8]

>>> reflect tree2
Node 4 (Node 7 (Node 8 Leaf Leaf) (Node 5 (Node 6 Leaf Leaf) Leaf)) (Node 2 (Node 3 Leaf Leaf) (Node 1 Leaf Leaf))

>>> inorder (reflect tree2)
[8,7,6,5,4,3,2,1]

height :: BinTree a -> Integer

The height function returns the height of a BinTree.

For example:

>>> tree1
Leaf

>>> height tree1
0

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))    

>>> height tree2
4

depth :: Ord a => a -> BinTree a -> Integer -> Integer

The depth function returns the depth of a node in a BinTree.

For example:

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))

>>> depth 1 tree2 0
2

>>> depth 2 tree2 0
1

>>> depth 4 tree2 0
0

>>> depth 8 tree2 0
2

size :: BinTree a -> Int

The size function returns the number of nodes in a BinTree. Leaf nodes are not counted.

For example:

>>> tree1
Leaf

>>> size tree1
0

>>> tree2
Node 4 (Node 2 (Node 1 Leaf Leaf) (Node 3 Leaf Leaf)) (Node 7 (Node 5 Leaf (Node 6 Leaf Leaf)) (Node 8 Leaf Leaf))

>>> size tree2
8

isBalanced :: BinTree a -> Bool

The isBalanced function checks is a BinTree is properly balanced.

For example:

>>> balanced tree2
False

>>> balanced tree1
True