Base

Description

Functionality for: * Converting decimal numbers to binary, octal, hexadecimal or any other base/radix * Converting numbers from binary, octal, hexadecimal or any other baseradix to decimal baseradix

Documentation

toBase :: Integer -> Integer -> [Integer]

The toBase function converts a number from decimal base to a specified base in the form of digits. The function takes two arguments, both of type Integer.

Below are a few examples:

>>> toBase 8 37
[4,5]

>>> toBase 100 233243
[23,32,43]

>>> toBase 9 233243
[3,8,4,8,4,8]

>>> toBase 35 233243
[5,15,14,3]

toBin :: Integer -> [Integer]

The toBin function converts a decimal number to its binary equivalent. The function takes one argument of type Integer, which is the decimal number.

Below are a few examples:

>>> toBin 11
	[1,0,1,1]

>>> toBin 32
[1,0,0,0,0,0]

>>> toBin (3^(-1))
*** Exception: Negative exponent

toOct :: Integer -> [Integer]

The toOct function converts a decimal number to its octal equivalent. The function takes one argument of type Integer, which is the decimal number.

Below are a few examples:

>>> toOct 11
	[1,3]

>>> toOct 100
[1,4,4]

toDec :: Integer -> [Integer]

The toDec function converts a decimal number to its decimal equivalent. The function returns a list of digits of the decimal number. The function takes one argument of type Integer, which is the decimal number.

Below are a few examples:

>>> toDec 15
	[1,5]

>>> toDec 123
[1,2,3]

toHex :: Integer -> [Integer]

The toHex function converts a decimal number to its hexadecimal equivalent. The function takes one argument of type Integer, which is the decimal number.

Below are a few examples:

>>> toHex 15
	[15]

>>> toHex 1200
[4,11,0]

fromBase :: Integer -> [Integer] -> Integer

The fromBase function converts a number from a specified base to decimal in the form of a list of digits. The function takes two arguments, the first of type Integer and the second of type [Integer].

Below are a few examples:

>>> fromBase 16 [15,15,15]
4095

>>> fromBase 100 [21,12,1]
211201

>>> fromBase 2 [1,0,1]
5

fromBin :: [Integer] -> Integer

The fromBin function converts a number from binary (in the form of a list of digits) to decimal. The function takes two arguments, the first of type Integer and the second of type [Integer].

Below are a few examples:

>>> fromBin [1,0,1,0]
	10

>>> fromBin (toBin 12345)
12345

fromOct :: [Integer] -> Integer

The fromOct function converts a number from octal (in the form of a list of digits) to decimal. The function takes two arguments, the first of type Integer and the second of type [Integer].

Below are a few examples:

>>> fromOct [6,3,7]
	415

>>> fromOct (toOct 1234)
1234

fromDec :: [Integer] -> Integer

The fromDec function converts a number from decimal (in the form of a list of digits) to decimal. The function takes two arguments, the first of type Integer and the second of type [Integer].

Below are a few examples:

>>> fromDec [1,5,8,2,2]
	15822

>>> fromDec (toDec 635465)
635465

fromHex :: [Integer] -> Integer

The fromHex function converts a number from hexadecimal (in the form of a list of digits) to decimal. The function takes two arguments, the first of type Integer and the second of type [Integer].

Below are a few examples:

>>> fromHex [14,15,8,2,0]
	981024

>>> fromHex (toHex 0234)
234

toAlphaDigits :: [Integer] -> String

The toAlphaDigits function converts a number from a list of Integer to corresponding String representation. The function takes one argument of type [Integer], which contains the list of digits.

Below are a few examples:

>>> toAlphaDigits $ toBase 16 23432
"5b88"

>>> toAlphaDigits $ toBase 16 255
"ff"

>>> toAlphaDigits [38,12,1]
"}c1"

>>> toAlphaDigits [21,12,1]
"lc1"

fromAlphaDigits :: String -> [Integer]

The 'from AlphaDigits' function converts a String to list of Integer digits. The function takes only one argument of type String.

Below are a few examples:

>>> fromAlphaDigits "j43hbrh"
[19,4,3,17,11,27,17]

>>> fromAlphaDigits "ffff"
[15,15,15,15]